Still have the bug of the kernel floats32 not loading correctly.

+++ cpp/MultiDimArray.h
@@ -14,12 +14,12 @@ class MultiDimArray {
     public:
 
     explicit MultiDimArray(std::vector<int> dims);
-    std::vector<double> values;
+    std::vector<float> values;
     std::vector<int> shape;
-    double* get(int dim0);
-    double* get(int dim0, int dim1);
-    double* get(int dim0, int dim1, int dim2);
-    double* get(int dim0, int dim1, int dim2, int dim3);
+    float* get(int dim0);
+    float* get(int dim0, int dim1);
+    float* get(int dim0, int dim1, int dim2);
+    float* get(int dim0, int dim1, int dim2, int dim3);
 };
